requirements.txt
pandas==0.24.1
Pillow==6.1.0
scikit_learn==0.21.3
